開発 #32139
開発体制強化 > Webデータベース > Database.commit複数処理

+++ portlets/gpdb/src/main/java/com/aimluck/eip/gpdb/GpdbKubunMultiDelete.java
@@ -133,8 +133,12 @@ public class GpdbKubunMultiDelete extends ALAbstractCheckList {
       // 削除処理
       for (EipMGpdbKubunValue value : flist) {
 
-        Database.delete(value);
-        Database.commit();
+        // 区分値を持たない区分があるかを調べ、削除。その際にcommitも行われるので、それ以外の時には別にcommitを行う。
+        if (!GpdbUtils.removeGpdbKubunNoValue(value.getGpdbKubunId(), value
+          .getGpdbKubunValueId(), value.getGpdbKubunValueId())) {
+          // 区分値を削除
+          Database.delete(value);
+        }
 
         EipMGpdbKubun Gpdbkubun =
           GpdbUtils.getEipMGpdbKubun(value.getGpdbKubunId());
@@ -148,11 +152,6 @@ public class GpdbKubunMultiDelete extends ALAbstractCheckList {
             + "」 値 「"
             + value.getGpdbKubunValue()
             + "」 削除");
-      }
-
-      // 区分値を持たない区分を削除
-      if (GpdbUtils.removeGpdbKubunNoValue()) {
-        Database.commit();
       }
 
     } catch (Exception ex) {
